% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{adduct_calc}
\alias{adduct_calc}
\title{TBA}
\usage{
adduct_calc(myadduct, DF)
}
\arguments{
\item{myadduct}{An adduct name , e.g. '[M+H]+' or '[M-H]-'
Complete list: [M+H]+ [M+Na]+ [M+NH4]+ [M+H-H2O]+ [M+2H]2+ [M+3H]3+ [M+4H]4+ 
[M+K]+ [M+2K]2+ [M+2Na]2+ [M+2Na-H]+ [M+2K-H]+ [M+Ag]+ [M+Li]+ 
[M+2Li]2+ [M+H+CH3CN]+ [M+Na+CH3CN]+ [M.NaFormate+H]+ [M.NH4Formate+H]+ 
[M-H]- [M-H-H2O]- [M-CH3]- [M-2H]2- [M-3H]3- [M-4H]4- [M.Cl]- [M.OAc]- 
[M+Na-2H]- [M+K-2H]- [M.Formate]- [M.NaFormate-H]- [M.NH4Formate-H]- 
[M.F]- [M.HF2]2-}

\item{DF}{The data frame generated by RefMet mapping, e.g. RefMet_mapped
The DF must contain the neutral mass column labeled 'Exact.mass'}
}
\value{
A data frame
}
\description{
The adduct_calc function.
}
\examples{
library(RefMet)

RefMet_mapped_with_adduct_col<- adduct_calc('[M+H]+',RefMet_mapped)
head(RefMet_mapped_with_adduct_col)

}
