% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetPlotPrint.R
\name{metplot}
\alias{metplot}
\title{Grab mass-spectrometry data from file(s)}
\usage{
metplot(mycat, metadata)
}
\arguments{
\item{mycat}{One of the following 6 metadata items:
'Species','Sample source','Disease association',
'Analysis type','MS polarity','Chromatography'}

\item{metadata}{One of the following 6 metadata items:
'Species','Sample source','Disease association',
'Analysis type','MS polarity','Chromatography'}
}
\value{
A ploteach named after the arguments requested in
  grab_what. E.g. $MS1 contains MS1 information, $MS2 contains fragmentation
  info, etc. MS1 data has four columns: retention time (rt), mass-to-charge
  (mz), intensity (int), and filename. MS2 data has six: retention time (rt),
}
\description{
The main `RaMS` function. This function accepts a list of the files that will
be read into R's working memory and returns a list of `data.table`s
if they are encoded in seconds.
}
\examples{
library(RefMet)

metadata<- refmet_metadata("Tyrosine")
head(metadata)
metplot('Species', metadata)

}
