% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetPlotPrint.R
\name{metprint}
\alias{metprint}
\title{TBA}
\usage{
metprint(mycat, metadata)
}
\arguments{
\item{mycat}{A metadata category name}

\item{metadata}{The metadata table for a metabolite of interest}
}
\value{
A file of results
}
\description{
The metprint function.
}
\examples{
library(RefMet)

metadata<- refmet_metadata("Tyrosine")
head(metadata)
metprint('Species',metadata)

}
