% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMet-package.R
\docType{package}
\name{RefMet-package}
\alias{RefMet}
\alias{RefMet-package}
\title{RefMet: RefMet:Tools for standardization of metabolite names}
\description{
R-based access to the RefMet metabolite name standardization system.
}
\details{
The main objective of RefMet (https://www.metabolomicsworkbench.org/databases/refmet/index.php) is to provide a standardized reference nomenclature
for both discrete metabolite structures and metabolite species identified by spectroscopic techniques in metabolomics experiments.
The wide range of available analytical methods coupled with the even wider range of metabolite databases and in-house libraries
has unfortunately led to a pervasive problem wherein the same metabolite species may be reported by many different names.
This nomenclature issue represents a significant barrier for comparative analysis of metabolomics data across studies generated by different institutions and/or platforms.
We have used data from over 3,000 MS and NMR studies on the Metabolomics Workbench to develop a highly curated analytical chemistry-centric
list of common names for metabolite structures and isobaric species.
This nomenclature is also linked to a metabolite classification system.
A name-conversion user interface is provided where users can submit a list of metabolite names and map them to the corresponding RefMet names.
RefMet has the ability to greatly increase the data-sharing potential of metabolomics experiments and facilitate "meta-analysis"
and systems biology objectives for the majority of commonly encountered metabolite species.
These R-based tools enable users to map a list of metabolite names (from a text file or a data frame column) to RefMet nomenclature in batch mode,
gain access to RefMet metadata for metabolites and perform plotting and saving functions.

Reference:RefMet: a reference nomenclature for metabolomics. Eoin Fahy and Shankar Subramaniam, Nature Methods, 17, 1173-1174 (2020)

See the package intro on GitHub at https://github.com/eoindfahy/RefMet for information on usage and examples
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eoindfahy/RefMet}
  \item Report bugs at \url{https://github.com/eoindfahy/RefMet/issues}
}

}
\author{
\strong{Maintainer}: Eoin Fahy \email{efahy@ucsd.edu} [copyright holder]

}
\keyword{internal}
