% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_map}
\alias{refmet_map}
\title{Map metabolite names to RefMet standardized nomenclature}
\usage{
refmet_map(thefile)
}
\arguments{
\item{thefile}{A file name for a text file of metabolite names (one per line)}
}
\value{
A data frame data frame containing 14 columns (input name, RefMet name, formula,
exact mass, super class, main class, sub class, PubChem CID, KEGG ID, ChEBI ID, HMDB ID,
LIPID MAPS ID InChIKey and RefMet_ID).
}
\description{
This function accepts a text file of metabolite names (one per line) as input.
The file contents are sent to the Metabolomics Workbench server where the
metabolite names are matched to RefMet standardized nomenclature (where possible).
The output is a data frame containing 14 columns (input name, RefMet name, formula,
exact mass, super class, main class, sub class, PubChem CID, KEGG ID, ChEBI ID, HMDB ID,
LIPID MAPS ID InChIKey and RefMet_ID).
}
\examples{
library(RefMet)
infile <- system.file("extdata", "met_list.txt", package = "RefMet")
#Use a text file of metabolite names as input
RefMet_mapped <- refmet_map(infile)
head(RefMet_mapped)

}
