% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_map_df}
\alias{refmet_map_df}
\title{Grab mass-spectrometry data from file(s)}
\usage{
refmet_map_df(DX)
}
\arguments{
\item{thefile}{A character vector of filenames to read into R's memory. Both
absolute and relative paths are acceptable.}
}
\value{
A data frame each named after the arguments requested in
  grab_what. E.g. $MS1 contains MS1 information, $MS2 contains fragmentation
  info, etc. MS1 data has four columns: retention time (rt), mass-to-charge
  (mz), intensity (int), and filename. MS2 data has six: retention time (rt),
}
\description{
The main `RaMS` function. This function accepts a list of the files that will
be read into R's working memory and returns a list of `data.table`s
if they are encoded in seconds.
}
\examples{
library(RefMet)

DF <- read.table(thefile, header = TRUE,  quote = "", sep="\t");
#Use a data frame column as input
RefMet_mapped <- refmet_map_df(DF[,1])
head(RefMet_mapped)

}
