% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_metadata}
\alias{refmet_metadata}
\title{Grab mass-spectrometry data from file(s)}
\usage{
refmet_metadata(metabolite)
}
\arguments{
\item{thefile}{A character vector of filenames to read into R's memory. Both
absolute and relative paths are acceptable.}
}
\value{
A data frame each named after the arguments requested in
  grab_what. E.g. $MS1 contains MS1 information, $MS2 contains fragmentation
  info, etc. MS1 data has four columns: retention time (rt), mass-to-charge
  (mz), intensity (int), and filename. MS2 data has six: retention time (rt),
}
\description{
The main `RaMS` function. This function accepts a list of the files that will
be read into R's working memory and returns a list of `data.table`s
if they are encoded in seconds.
}
\examples{
library(RefMet)

metadata<- refmet_metadata("Tyrosine")
head(metadata)

}
