% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetPlotPrint.R
\name{classplot}
\alias{classplot}
\title{Grab mass-spectrometry data from file(s)}
\usage{
classplot(mycat, RefMet_mapped)
}
\arguments{
\item{mycat, }{One of the following 6 metadata items:
'Species','Sample source','Disease association',
'Analysis type','MS polarity','Chromatography'}

\item{RefMet_mapped, }{One of the following 6 metadata items:
'Species','Sample source','Disease association',
'Analysis type','MS polarity','Chromatography'}
}
\value{
A plot each named after the arguments requested in
  grab_what. E.g. $MS1 contains MS1 information, $MS2 contains fragmentation
  info, etc. MS1 data has four columns: retention time (rt), mass-to-charge
  (mz), intensity (int), and filename. MS2 data has six: retention time (rt),
}
\description{
The main `RaMS` function. This function accepts a list of the files that will
Retention times are reported in minutes, and will be converted automatically
if they are encoded in seconds.
}
\examples{
library(RefMet)

infile <- system.file("extdata", "met_list.txt", package = "RefMet")
#Use a text file of metabolite names as input
RefMet_mapped <- refmet_map(infile)
classplot('Main class',RefMet_mapped)

}
