% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetPlotPrint.R
\name{metplot}
\alias{metplot}
\title{TBA}
\usage{
metplot(mycat, metadata, pwidth = 15, pheight = 8)
}
\arguments{
\item{mycat}{A metadata category name}

\item{metadata}{The metadata table for a metabolite of interest}
}
\value{
A plot
}
\description{
The metplot function.
}
\examples{
library(RefMet)

metadata<- refmet_metadata("Tyrosine")
head(metadata)
metplot('Species', metadata)

}
