% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMet-package.R
\docType{package}
\name{RefMet-package}
\alias{RefMet}
\alias{RefMet-package}
\title{RefMet: RefMet:Tools for standardization of metabolite names}
\description{
R-based access to the RefMet metabolite name standardization system.
}
\details{
to be added

See the package intro on GitHub at https://github.com/eoindfahy/RefMet
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eoindfahy/RefMet}
  \item Report bugs at \url{https://github.com/eoindfahy/RefMet/issues}
}

}
\author{
\strong{Maintainer}: Eoin Fahy \email{efahy@ucsd.edu} [copyright holder]

}
\keyword{internal}
