% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_map}
\alias{refmet_map}
\title{TBA}
\usage{
refmet_map(thefile)
}
\arguments{
\item{thefile}{A file name}
}
\value{
A data frame
}
\description{
The refmet_map function.
}
\examples{
library(RefMet)
infile <- system.file("extdata", "met_list.txt", package = "RefMet")
#Use a text file of metabolite names as input
RefMet_mapped <- refmet_map(infile)
head(RefMet_mapped)

}
