% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_map_df}
\alias{refmet_map_df}
\title{Grab mass-spectrometry data from file(s)}
\usage{
refmet_map_df(DX)
}
\arguments{
\item{DX}{(a data frame column with metabolite names}
}
\value{
A data frame  (rt),
}
\description{
The refmet_map_df function.
}
\examples{
library(RefMet)

DF <- read.table(thefile, header = TRUE,  quote = "", sep="\t");
#Use a data frame column as input
RefMet_mapped <- refmet_map_df(DF[,1])
head(RefMet_mapped)

}
