% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_map_df}
\alias{refmet_map_df}
\title{Map metabolite names to RefMet standardized nomenclature}
\usage{
refmet_map_df(DX)
}
\arguments{
\item{DX}{A data frame column of metabolite names}
}
\value{
A data frame data frame containing 7 columns (input name, RefMet name, formula,
exact mass, super class, main class and sub class).
}
\description{
This function accepts a data frame column of metabolite names.
The column contents are sent to the Metabolomics Workbench server where the
metabolite names are matched to RefMet standardized nomenclature (where possible).
The output is a data frame containing 7 columns (input name, RefMet name, formula,
exact mass, super class, main class and sub class).
}
\examples{
library(RefMet)

DF <- read.table(thefile, header = TRUE,  quote = "", sep="\t");
#Use a data frame column as input
RefMet_mapped <- refmet_map_df(DF[,1])
head(RefMet_mapped)

}
