% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMetCode.R
\name{refmet_metadata}
\alias{refmet_metadata}
\title{TBA}
\usage{
refmet_metadata(metabolite)
}
\arguments{
\item{metabolite}{A metabolite name}
}
\value{
A data frame
}
\description{
The refmet_map function.
}
\examples{
library(RefMet)

metadata<- refmet_metadata("Tyrosine")
head(metadata)

}
